Decoupl and refactor ViolenceCases to SemanticGraph component for visualizing relationships and interactions in data
